紧急修复,修复了删除节点的时候，删除了有数据的表。

+++ Components/BP.En30/Sys/FrmRpt.cs
@@ -1205,7 +1205,8 @@ namespace BP.Sys
             DBAccess.RunSQLs(sql);
             try
             {
-                BP.DA.DBAccess.RunSQL("DROP TABLE " + this.PTable);
+                if (DBAccess.RunSQLReturnValInt("SELECT COUNT(*) as Num FROM " + this.PTable + " WHERE 1=1 ") == 0)
+                    BP.DA.DBAccess.RunSQL("DROP TABLE " + this.PTable);
             }
             catch
             {
